added FizzBuzz with UnitTest
FizzBuzz is one popular Interview Question, hence I created a possible
solution using modulo - The return is an Array with Numbers Fizz Buzz
and FizzBuzz Strings.

FizzBuzzTest is checking this array wether all Fizz Buzz and FizzBuzzes
are at the right position. At the moment we don't check for numbers.