added ReverseString with UnitTest
ReverseString will reverse a input String using an character array,
if the input String contains a new line or a null character at the
_end_ of the String, it will not be reversed (put at the front).
The null character will be replaced with a '\n' and put at the
end of the returning String. A new line will just stay at the end
of the output String.